Done server implementation. Test are left.
