+++ src/PrimesServerApplication.cpp
@@ -12,7 +12,7 @@ int main(int argc, char *argv[]) {
         return 0;
     }
     if (argc == 2) {
-        uint16_t portno = static_cast<uint16_t>(strtol(argv[2], nullptr, 10));
+        auto portno = static_cast<uint16_t>(strtol(argv[1], nullptr, 10));
         PrimesServerApplication(portno).run();
     } else {
         PrimesServerApplication().run();
@@ -24,11 +24,10 @@ int main(int argc, char *argv[]) {
 void PrimesServerApplication::run() {
     printUsage();
 
-    std::thread server_thread(&Server::run, server);
+    std::thread server_thread(&Server::run, std::ref(server));
     bool toExit = false;
     while (!toExit) {
         std::string command;
-        std::cout << ">";
         std::cin >> command;
         if (command == "shutdown") {
             uint32_t client_id;
@@ -36,23 +35,26 @@ void PrimesServerApplication::run() {
             if (server.shutdown_client(client_id)) {
                 std::cout << "Client was disconnected successfully\n";
             } else {
-                std::cout << "Client was unsuccessfully. Are you sure this client exists?\n";
+                std::cout << "No such client\n";
             }
         } else if (command == "exit") {
+            server.stop();
             toExit = true;
         } else {
             std::cout << "UNKNOWN COMMAND\n";
         }
     }
+    server_thread.join();
 }
 
 void PrimesServerApplication::printUsage() {
-    std::cout << "Commands:\n"
+    std::cout << "this server runs on " << portno << " port\n"
+    << "Commands:\n"
     << "shutdown <client_id> - close connection with specific client\n"
     << "exit - close server\n"
     << "\n";
 }
 
-PrimesServerApplication::PrimesServerApplication(uint16_t portno) : portno(portno) {}
+PrimesServerApplication::PrimesServerApplication(uint16_t portno) : portno(portno), server(portno, 10) {}
 
 PrimesServerApplication::PrimesServerApplication() : PrimesServerApplication(PrimesServerApplication::DEFAULT_PORTNO) {}
